Moves color functions to separate functions for ease of testing
